Refactor the basket to a separate component
