+++ src/components/Header.tsx
@@ -1,12 +1,9 @@
 import * as React from 'react';
 import { connect } from 'react-redux';
 import { MyStore, getRaster, getWMS } from '../reducers';
-import { removeRasterFromBasket, removeWMSFromBasket } from '../action';
-import { Raster, LatLng, Bootstrap, WMS } from '../interface';
+import { Raster, Bootstrap, WMS } from '../interface';
+import Basket from './Basket';
 import './styles/Header.css';
-
-import { zoomLevelCalculation, getCenterPoint } from '../utils/latLngZoomCalculation';
-import { openAllInLizard } from '../utils/url';
 
 interface MyProps {
     showProfileDropdown: boolean,
@@ -19,34 +16,7 @@ interface PropsFromState {
     user: Bootstrap['user']
 };
 
-interface PropsFromDispatch {
-    removeRasterFromBasket: (uuid: string) => void,
-    removeWMSFromBasket: (uuid: string) => void
-};
-
-type HeaderProps = MyProps & PropsFromState & PropsFromDispatch;
-
-class Header extends React.Component<HeaderProps> {
-    //Open All Data button will open all rasters and WMS layers in Lizard Client
-    //with projection of the last selected raster
-    openInLizard = (rasters: Raster[], wmsLayers: WMS[]) => {
-        //Get the last selected raster in the basket which is the first item in the rasters array
-        const lastSelectedRaster = rasters.length !== 0 ? rasters[0] : null;
-
-        //Get the spatial bounds of the last selected raster, if spatial_bounds is null then set it to the global map
-        const { north, east, south, west } = (lastSelectedRaster && lastSelectedRaster.spatial_bounds) ?
-            lastSelectedRaster.spatial_bounds : { north: 85, east: 180, south: -85, west: -180 };
-
-        //Get the center point of the raster based on its spatial bounds
-        const centerPoint: LatLng = getCenterPoint(north, east, south, west);
-
-        //Calculate the zoom level of the last selected raster by using the zoomLevelCalculation function
-        const zoom = zoomLevelCalculation(north, east, south, west);
-
-        //Open all rasters and WMS layers in Lizard
-        openAllInLizard(rasters, centerPoint, zoom, wmsLayers);
-    };
-
+class Header extends React.Component<MyProps & PropsFromState> {
     renderProfileDropdown() {
         return (
             <div className="user-profile_dropdown" onMouseLeave={this.props.toggleProfileDropdown}>
@@ -67,7 +37,7 @@ class Header extends React.Component<HeaderProps> {
     };
 
     render() {
-        const { rasters, wmsLayers , removeRasterFromBasket, removeWMSFromBasket, user } = this.props;
+        const { rasters, wmsLayers, user } = this.props;
         const basket = [...rasters, ...wmsLayers];
 
         return (
@@ -107,57 +77,12 @@ class Header extends React.Component<HeaderProps> {
                     </a>
                 </div>
                 {/*This is the PopUp window when the basket is clicked*/}
-                <div className="header-popup" id="basket">
-                    <div className="header-popup__content">
-                        <h3>My selection</h3>
-                        <div className="header-popup__content-bovenste-laag">
-                            <span className="header-popup__content-item">{basket.length} items</span>
-                            <span className="header-popup__content-layer-title">Upper layer</span>
-                        </div>
-                        <ul className="header-popup__content-ul">
-                            {rasters.map(raster => (
-                                <li className="header-popup__content-li" key={raster.uuid}>
-                                    {raster.temporal ?
-                                        <img className="li li-type" src="image/raster-temporal.svg" alt="raster" /> :
-                                        <img className="li li-type" src="image/raster-non-temporal.svg" alt="raster" />
-                                    }
-                                    <div className="li li-name">{raster.name}</div>
-                                    <div className="li li-org">{raster.organisation && raster.organisation.name}</div>
-                                    {/* <div className="li li-obs">{raster.observation_type && raster.observation_type.parameter}</div> */}
-                                    <div className="li li-time">{new Date(raster.last_modified).toLocaleDateString()}</div>
-                                    <div className="li li-basket li-basket__icon-box" onClick={() => removeRasterFromBasket(raster.uuid)}>
-                                        <svg className="li-basket__icon">
-                                            <use xlinkHref="image/symbols.svg#icon-remove_shopping_cart" />
-                                        </svg>
-                                    </div>
-                                </li>
-                            ))}
-                            {wmsLayers.map(wms => (
-                                <li className="header-popup__content-li" key={wms.uuid}>
-                                    <img className="li li-type" src="image/wms.svg" alt="wms" />
-                                    <div className="li li-name">{wms.name}</div>
-                                    <div className="li li-org">{wms.organisation && wms.organisation.name}</div>
-                                    {/* <div className="li li-obs">{wms.slug}</div> */}
-                                    <div className="li li-time" />
-                                    <div className="li li-basket li-basket__icon-box" onClick={() => removeWMSFromBasket(wms.uuid)}>
-                                        <svg className="li-basket__icon">
-                                            <use xlinkHref="image/symbols.svg#icon-remove_shopping_cart" />
-                                        </svg>
-                                    </div>
-                                </li>
-                            ))}
-                        </ul>
-                        <p className="header-popup__content-onderste-laag">Lower layer</p>
-                        <button
-                            className="header-popup__content-button"
-                            disabled={basket.length === 0 ? true : false}
-                            onClick={() => this.openInLizard(rasters, wmsLayers)}
-                        >
-                            Open all data in Lizard
-                        </button>
-                        {/* eslint-disable-next-line */}
-                        <a href="#" className="header-popup__close">&times;</a>
-                    </div>
+                <div className="header-basket" id="basket">
+                    <Basket
+                        rasters={rasters}
+                        wmsLayers={wmsLayers}
+                        basket={basket}
+                    />
                 </div>
                 {/*This is the PopUp window for the Information box*/}
                 <div className="header-information-box" id="information">
@@ -197,9 +122,4 @@ const mapStateToProps = (state: MyStore): PropsFromState => {
     };
 };
 
-const mapDispatchToProps = (dispatch): PropsFromDispatch => ({
-    removeRasterFromBasket: (uuid: string) => removeRasterFromBasket(uuid, dispatch),
-    removeWMSFromBasket: (uuid: string) => removeWMSFromBasket(uuid, dispatch)
-});
-
-export default connect(mapStateToProps, mapDispatchToProps)(Header);
+export default connect(mapStateToProps)(Header);
